requirements.txt
allure-pytest==2.9.45
pytest==7.0.1
selenium==4.1.0
Faker